refactor: avoid need for a plugin:types command
